I did something, but who knows what...

+++ app/src/main/java/com/example/strawberry_app/network/ConnectionState.kt
@@ -5,5 +5,5 @@ sealed class ConnectionState {
     object Connecting : ConnectionState()
     object Disconnected : ConnectionState()
     data class Error(val message: String, val canRetry: Boolean = true) : ConnectionState()
-    data class Reconnecting(val attempt: Int, val time: Long): ConnectionState()
+    data class Reconnecting(val attempt: Long, val time: Long): ConnectionState()
 }
